# docs/openapi/repository-api-v1.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "description": "Error detail object for the public API error envelope.",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured metadata to help clients debug the error.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable summary of the failure.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "Standardized error envelope shared across HTTP endpoints.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "Normalized error payload."
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "SkillSearchResponse": {
        "description": "Compact advisory search response.",
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SkillSearchResultResponse"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "SkillSearchResponse",
        "type": "object"
      },
      "SkillSearchResultResponse": {
        "description": "Compact advisory candidate returned by the search API.",
        "properties": {
          "content_size_bytes": {
            "title": "Content Size Bytes",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "freshness_days": {
            "title": "Freshness Days",
            "type": "integer"
          },
          "lifecycle_status": {
            "enum": [
              "published",
              "deprecated",
              "archived"
            ],
            "title": "Lifecycle Status",
            "type": "string"
          },
          "matched_fields": {
            "items": {
              "type": "string"
            },
            "title": "Matched Fields",
            "type": "array"
          },
          "matched_tags": {
            "items": {
              "type": "string"
            },
            "title": "Matched Tags",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "trust_tier": {
            "enum": [
              "untrusted",
              "internal",
              "verified"
            ],
            "title": "Trust Tier",
            "type": "string"
          },
          "usage_count": {
            "title": "Usage Count",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "version",
          "name",
          "description",
          "tags",
          "lifecycle_status",
          "trust_tier",
          "published_at",
          "freshness_days",
          "content_size_bytes",
          "usage_count",
          "matched_fields",
          "matched_tags",
          "reasons"
        ],
        "title": "SkillSearchResultResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Registry-first API for immutable skill publication, indexed search candidate\nretrieval, direct relationship reads, exact version retrieval, artifact streaming,\nand version listing.\n\nThe server owns data-local registry operations only. Prompt interpretation,\nreranking, dependency solving, final selection, lock generation, and execution\nplanning remain client-owned behavior outside this API boundary.",
    "title": "aptitude-server Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/discovery/skills/search": {
      "get": {
        "description": "Return compact advisory candidates from indexed metadata and description search. This route is discovery-only: it does not resolve relationships, choose a final candidate, or perform dependency solving.",
        "operationId": "searchDiscoverySkills",
        "parameters": [
          {
            "description": "Full-text discovery query.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Full-text discovery query.",
              "title": "Q"
            }
          },
          {
            "description": "Repeatable tag filter; all values must match.",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeatable tag filter; all values must match.",
              "title": "Tag"
            }
          },
          {
            "description": "Language alias filter implemented as a normalized tag.",
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Language alias filter implemented as a normalized tag.",
              "title": "Language"
            }
          },
          {
            "description": "Maximum age in days since publication.",
            "in": "query",
            "name": "fresh_within_days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum age in days since publication.",
              "title": "Fresh Within Days"
            }
          },
          {
            "description": "Maximum allowed markdown size in bytes.",
            "in": "query",
            "name": "max_content_size_bytes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum allowed markdown size in bytes.",
              "title": "Max Content Size Bytes"
            }
          },
          {
            "description": "Repeatable lifecycle-state filter.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "published",
                      "deprecated",
                      "archived"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeatable lifecycle-state filter.",
              "title": "Status"
            }
          },
          {
            "description": "Repeatable trust-tier filter.",
            "in": "query",
            "name": "trust_tier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "untrusted",
                      "internal",
                      "verified"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeatable trust-tier filter.",
              "title": "Trust Tier"
            }
          },
          {
            "description": "Maximum number of skill candidates to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Maximum number of skill candidates to return.",
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "results": [
                    {
                      "content_size_bytes": 123,
                      "description": "Linting skill",
                      "freshness_days": 0,
                      "lifecycle_status": "published",
                      "matched_fields": [
                        "name",
                        "tags"
                      ],
                      "matched_tags": [
                        "python"
                      ],
                      "name": "Python Lint",
                      "published_at": "2026-03-10T08:30:00Z",
                      "reasons": [
                        "text_match",
                        "tag_filter_match"
                      ],
                      "slug": "python.lint",
                      "tags": [
                        "python",
                        "lint"
                      ],
                      "trust_tier": "internal",
                      "usage_count": 0,
                      "version": "1.2.3"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillSearchResponse"
                }
              }
            },
            "description": "Compact advisory skill candidates returned successfully."
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "invalid_request": {
                    "value": {
                      "error": {
                        "code": "INVALID_REQUEST",
                        "details": {
                          "errors": [
                            {
                              "input": "latest",
                              "loc": [
                                "path",
                                "version"
                              ],
                              "msg": "String should match pattern",
                              "type": "string_pattern_mismatch"
                            }
                          ]
                        },
                        "message": "Request validation failed."
                      }
                    }
                  },
                  "missing_selector": {
                    "value": {
                      "error": {
                        "code": "INVALID_REQUEST",
                        "details": {
                          "errors": [
                            {
                              "input": {
                                "limit": 20,
                                "status": [],
                                "tags": [],
                                "trust_tier": []
                              },
                              "loc": [],
                              "msg": "Value error, At least one search selector must be provided.",
                              "type": "value_error"
                            }
                          ]
                        },
                        "message": "Search request validation failed."
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The discovery request is invalid."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search skill discovery metadata",
        "tags": [
          "discovery"
        ]
      }
    },
...
  }
}